updated to uze shared profile router
