+++ lib/api.js
@@ -34,7 +34,6 @@ module.exports = settings => {
   app.use(require('./middleware/user'));
 
   app.use('/me', require('./routers/user'));
-
   app.use('/establishment(s)?', require('./routers/establishment'));
 
   app.use((req, res, next) => {
