+++ lib/middleware/user.js
@@ -18,15 +18,4 @@ router.use((req, res, next) => {
     .catch(next);
 });
 
-router.use((req, res, next) => {
-  Promise.resolve()
-    .then(() => req.user.can())
-    .then(response => response.json)
-    .then(allowedActions => {
-      req.allowedActions = allowedActions;
-    })
-    .then(() => next())
-    .catch(next);
-});
-
 module.exports = router;
