+++ lib/routers/profile/index.js
@@ -1,22 +1,49 @@
+const { omit } = require('lodash');
 const { Router } = require('express');
 const isUUID = require('uuid-validate');
+const { validateSchema } = require('../../middleware');
 const { NotFoundError } = require('../../errors');
 
-const router = Router({ mergeParams: true });
+const submit = (action) => {
+  return (req, res, next) => {
+    const params = {
+      action,
+      model: 'profile',
+      data: { ...req.body },
+      id: req.profileId
+    };
 
-router.param('id', (req, res, next, id) => {
-  if (!isUUID(id)) {
-    return next(new NotFoundError());
+    req.workflow(params)
+      .then(response => {
+        res.response = response;
+        next();
+      })
+      .catch(next);
+  };
+};
+
+const validate = () => {
+  return (req, res, next) => {
+    const ignoredFields = ['comments'];
+    return validateSchema(req.models.Profile, {
+      ...(req.profile || {}),
+      ...omit(req.body, ignoredFields)
+    })(req, res, next);
+  };
+};
+
+const getSingleProfile = req => {
+  if (!isUUID(req.profileId)) {
+    throw new NotFoundError();
   }
   const { Profile } = req.models;
-
   const profile = Profile.scopeSingle({
-    id: req.params.id,
-    establishmentId: req.establishment.id,
-    userId: req.profile.id
+    id: req.profileId,
+    profileId: req.profile.id,
+    establishmentId: (req.establishment && req.establishment.id) || undefined
   });
 
-  Promise.resolve()
+  return Promise.resolve()
     .then(() => req.user.can('profile.read.all', req.params))
     .then(allowed => {
       if (allowed) {
@@ -35,19 +62,17 @@ router.param('id', (req, res, next, id) => {
       if (!profile) {
         throw new NotFoundError();
       }
-      res.profile = profile;
-      next();
-    })
-    .catch(next);
-});
+      return profile;
+    });
+};
 
-router.get('/', (req, res, next) => {
+const getAllProfiles = req => {
   const { Profile } = req.models;
   const { search, sort, filters, limit, offset } = req.query;
 
   const profiles = Profile.scopeToParams({
-    userId: req.profile.id,
-    establishmentId: req.establishment.id,
+    establishmentId: (req.establishment && req.establishment.id) || undefined,
+    profileId: req.profile.id,
     search,
     limit,
     offset,
@@ -55,7 +80,7 @@ router.get('/', (req, res, next) => {
     filters
   });
 
-  Promise.resolve()
+  return Promise.resolve()
     .then(() => req.user.can('profile.read.all', req.params))
     .then(allowed => {
       if (allowed) {
@@ -69,21 +94,36 @@ router.get('/', (req, res, next) => {
           }
           throw new NotFoundError();
         });
-    })
+    });
+};
+
+const router = Router({ mergeParams: true });
+
+router.use((req, res, next) => {
+  if (req.profileId) {
+    return Promise.resolve()
+      .then(() => getSingleProfile(req))
+      .then(profile => {
+        res.response = profile;
+        next();
+      })
+      .catch(next);
+  }
+  Promise.resolve()
+    .then(() => getAllProfiles(req))
     .then(({ filters, total, profiles }) => {
       res.meta.filters = filters;
       res.meta.total = total;
       res.meta.count = profiles.total;
       res.response = profiles.results;
-      return next();
+      next();
     })
     .catch(next);
 });
 
-router.get('/:id', (req, res, next) => {
-  res.response = res.profile;
-  next();
-});
+router.put('/', validate(), submit('update'));
+
+router.use('/training', require('./training-modules'));
 
 router.put('/:id',
   permissions('profile.update'),
